+++ ffi/tests/cpp-consumer/main.cpp
@@ -1,7 +1,5 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-#pragma once
-
-#include <tav/byte_buffer.h>
-#include <tav/errors.h>
+#define DOCTEST_CONFIG_IMPLEMENT_WITH_MAIN
+#include "doctest.h"